Added Justin's updated code, as well as AI pseudocode framework
The code we worked on today was added.  This mostly entailed getting AI
prepped in pseudocode and fixing a bug we were having with typecasting,
as well as ensuring that Dijkstra was up to standards.

+++ Strat of the ages/dsf/Map.cpp
@@ -6,5 +6,5 @@ WorldMap::WorldMap(){
 	for(int i = 0; i < 500; i++)\
 		for(int j = 0; j < 500; j++)
 			Prov = new Province(2,i+j*500,1.0f,1.0f);
-			Provinces.add((int)Prov);
+			Provinces.add(Prov);
 }
